Move grader generator to a plugin
