+++ inginious/frontend/webapp/pages/course_admin/task_edit.py
@@ -76,8 +76,6 @@ class CourseEditTask(INGIniousAdminPage):
                 task_data.get(
                     'problems',
                     {})),
-            json.dumps(
-                task_data.get('grader_test_cases', [])),
             self.contains_is_html(task_data),
             current_filetype,
             available_filetypes,
@@ -181,116 +179,6 @@ class CourseEditTask(INGIniousAdminPage):
             del problem_content["custom"]
 
         return problem_content
-
-    def parse_grader_test_case(self, test_case_content):
-        if not test_case_content["input_file"]:
-            raise Exception("Invalid input file in grader test case")
-
-        if not test_case_content["output_file"]:
-            raise Exception("Invalid output file in grader test case")
-
-        try:
-            test_case_content["weight"] = float(test_case_content["weight"])
-        except:
-            raise Exception("The weight for grader test cases must be a float")
-
-        test_case_content["diff_shown"] = "diff_shown" in test_case_content
-
-        return test_case_content
-
-    def preprocess_grader_data(self, data):
-        try:
-            data["grader_diff_max_lines"] = int(data["grader_diff_max_lines"])
-        except:
-            return json.dumps({"status": "error", "message": "'Maximum diff lines' must be an integer"})
-
-        if data["grader_diff_max_lines"] <= 0:
-            return json.dumps({"status": "error", "message": "'Maximum diff lines' must be positive"})
-
-        try:
-            data["grader_diff_context_lines"] = int(data["grader_diff_context_lines"])
-        except:
-            return json.dumps({"status": "error", "message": "'Diff context lines' must be an integer"})
-
-        if data["grader_diff_context_lines"] <= 0:
-            return json.dumps({"status": "error", "message": "'Diff context lines' must be positive"})
-
-        data["grader_compute_diffs"] = "grader_compute_diffs" in data
-        data["treat_non_zero_as_runtime_error"] = "treat_non_zero_as_runtime_error" in data
-        data["generate_grader"] = "generate_grader" in data
-
-        grader_test_cases = self.dict_from_prefix("grader_test_cases", data) or OrderedDict()
-
-        # Remove test-case dirty entries.
-        keys_to_remove = [key for key, _ in data.items() if key.startswith("grader_test_cases[")]
-        for key in keys_to_remove:
-            del data[key]
-
-        data["grader_test_cases"] = [self.parse_grader_test_case(val) for _, val in grader_test_cases.items()]
-        data["grader_test_cases"].sort(key=lambda test_case: (test_case["input_file"], test_case["output_file"]))
-
-        if len(set(test_case["input_file"] for test_case in data["grader_test_cases"])) != len(data["grader_test_cases"]):
-            return json.dumps({"status": "error", "message": "Duplicated input files in grader"})
-
-        return None
-
-    def postprocess_grader_data(self, data, task_fs):
-        for test_case in data["grader_test_cases"]:
-            if not task_fs.exists(test_case["input_file"]):
-                return json.dumps(
-                    {"status": "error", "message": "Grader input file does not exist: " + test_case["input_file"]})
-
-            if not task_fs.exists(test_case["output_file"]):
-                return json.dumps(
-                    {"status": "error", "message": "Grader output file does not exist: " + test_case["output_file"]})
-
-        if data["generate_grader"]:
-            if "grader_problem_id" not in data:
-                return json.dumps({"status": "error", "message": "Grader: the problem was not specified"})
-
-            if data["grader_problem_id"] not in data["problems"]:
-                return json.dumps({"status": "error", "message": "Grader: problem does not exist"})
-
-            problem_type = data["problems"][data["grader_problem_id"]]["type"]
-            if problem_type not in ['code-multiple-languages', 'code-file-multiple-languages']:
-                return json.dumps({"status": "error",
-                    "message": "Grader: only Code Multiple Language and Code File Multiple Language problems are supported"})
-
-            current_directory = os.path.dirname(__file__)
-            run_file_template_path = os.path.join(current_directory, '../../templates/course_admin/run_file_template.txt')
-            run_file_template = None
-
-            with open(run_file_template_path, "r") as f:
-                run_file_template = f.read()
-
-            problem_id = data["grader_problem_id"]
-            test_cases = [(test_case["input_file"], test_case["output_file"]) for test_case in data["grader_test_cases"]]
-            weights = [test_case["weight"] for test_case in data["grader_test_cases"]]
-            options = {
-                "compute_diff": data["grader_compute_diffs"],
-                "treat_non_zero_as_runtime_error": data["treat_non_zero_as_runtime_error"],
-                "diff_max_lines": data["grader_diff_max_lines"],
-                "diff_context_lines": data["grader_diff_context_lines"],
-                "output_diff_for": [test_case["input_file"] for test_case in data["grader_test_cases"]
-                    if test_case["diff_shown"]]
-            }
-
-            if len(test_cases) == 0:
-                return json.dumps(
-                    {"status": "error", "message": "You must provide test cases to autogenerate the grader"})
-
-            with tempfile.TemporaryDirectory() as temporary_folder_name:
-                run_file_name = 'run'
-                target_run_file = os.path.join(temporary_folder_name, run_file_name)
-
-                with open(target_run_file, "w") as f:
-                    f.write(run_file_template.format(
-                        problem_id=repr(problem_id), test_cases=repr(test_cases),
-                        options=repr(options), weights=repr(weights)))
-
-                task_fs.copy_to(temporary_folder_name)
-
-        return None
 
     def wipe_task(self, courseid, taskid):
         """ Wipe the data associated to the taskid from DB"""
@@ -429,10 +317,6 @@ class CourseEditTask(INGIniousAdminPage):
         task_fs = self.task_factory.get_task_fs(courseid, taskid)
         task_fs.ensure_exists()
 
-        error = self.preprocess_grader_data(data)
-        if error is not None:
-            return error
-
         # Call plugins and return the first error
         plugin_results = self.plugin_manager.call_hook('task_editor_submit', course=course, taskid=taskid,
                                                        task_data=data, task_fs=task_fs)
@@ -459,10 +343,6 @@ class CourseEditTask(INGIniousAdminPage):
                     return json.dumps(
                         {"status": "error", "message": "There was a problem while extracting the zip archive. Some files may have been modified"})
                 task_fs.copy_to(tmpdirname)
-
-        error = self.postprocess_grader_data(data, task_fs)
-        if error is not None:
-            return error
 
         self.task_factory.delete_all_possible_task_files(courseid, taskid)
         self.task_factory.update_task_descriptor_content(courseid, taskid, data, force_extension=file_ext)
