+++ inginious/frontend/webapp/plugins/grader_generator/__init__.py
@@ -1,0 +1,141 @@
+import json
+import tempfile
+import os.path
+from inginious.frontend.webapp.pages.course_admin.task_edit import CourseEditTask
+from collections import OrderedDict
+from inginious.frontend.webapp.plugins.utils import create_static_resource_page
+
+
+_PLUGIN_PATH = os.path.dirname(__file__)
+_BASE_RENDERER_PATH = _PLUGIN_PATH
+_BASE_STATIC_FOLDER = os.path.join(_PLUGIN_PATH, 'static')
+_RUN_FILE_TEMPLATE_PATH = os.path.join(_PLUGIN_PATH, 'run_file_template.txt')
+
+
+def parse_grader_test_case(self, test_case_content):
+    if not test_case_content["input_file"]:
+        raise Exception("Invalid input file in grader test case")
+
+    if not test_case_content["output_file"]:
+        raise Exception("Invalid output file in grader test case")
+
+    try:
+        test_case_content["weight"] = float(test_case_content["weight"])
+    except:
+        raise Exception("The weight for grader test cases must be a float")
+
+    test_case_content["diff_shown"] = "diff_shown" in test_case_content
+
+    return test_case_content
+
+
+def on_task_editor_submit(course, taskid, task_data, task_fs):
+    try:
+        task_data["grader_diff_max_lines"] = int(task_data["grader_diff_max_lines"])
+    except:
+        return json.dumps({"status": "error", "message": "'Maximum diff lines' must be an integer"})
+
+    if task_data["grader_diff_max_lines"] <= 0:
+        return json.dumps({"status": "error", "message": "'Maximum diff lines' must be positive"})
+
+    try:
+        task_data["grader_diff_context_lines"] = int(task_data["grader_diff_context_lines"])
+    except:
+        return json.dumps({"status": "error", "message": "'Diff context lines' must be an integer"})
+
+    if task_data["grader_diff_context_lines"] <= 0:
+        return json.dumps({"status": "error", "message": "'Diff context lines' must be positive"})
+
+    task_data["grader_compute_diffs"] = "grader_compute_diffs" in task_data
+    task_data["treat_non_zero_as_runtime_error"] = "treat_non_zero_as_runtime_error" in task_data
+    task_data["generate_grader"] = "generate_grader" in task_data
+
+    grader_test_cases = CourseEditTask.dict_from_prefix("grader_test_cases", task_data) or OrderedDict()
+
+    # Remove test-case dirty entries.
+    keys_to_remove = [key for key, _ in task_data.items() if key.startswith("grader_test_cases[")]
+    for key in keys_to_remove:
+        del task_data[key]
+
+    task_data["grader_test_cases"] = [parse_grader_test_case(val) for _, val in grader_test_cases.items()]
+    task_data["grader_test_cases"].sort(key=lambda test_case: (test_case["input_file"], test_case["output_file"]))
+
+    if len(set(test_case["input_file"] for test_case in task_data["grader_test_cases"])) != len(task_data["grader_test_cases"]):
+        return json.dumps({"status": "error", "message": "Duplicated input files in grader"})
+
+    for test_case in task_data["grader_test_cases"]:
+        if not task_fs.exists(test_case["input_file"]):
+            return json.dumps(
+                {"status": "error", "message": "Grader input file does not exist: " + test_case["input_file"]})
+
+        if not task_fs.exists(test_case["output_file"]):
+            return json.dumps(
+                {"status": "error", "message": "Grader output file does not exist: " + test_case["output_file"]})
+
+    if task_data["generate_grader"]:
+        if "grader_problem_id" not in task_data:
+            return json.dumps({"status": "error", "message": "Grader: the problem was not specified"})
+
+        if task_data["grader_problem_id"] not in task_data["problems"]:
+            return json.dumps({"status": "error", "message": "Grader: problem does not exist"})
+
+        problem_type = task_data["problems"][task_data["grader_problem_id"]]["type"]
+        if problem_type not in ['code-multiple-languages', 'code-file-multiple-languages']:
+            return json.dumps({"status": "error",
+                               "message": "Grader: only Code Multiple Language and Code File Multiple Language problems are supported"})
+
+        with open(_RUN_FILE_TEMPLATE_PATH, "r") as f:
+            run_file_template = f.read()
+
+        problem_id = task_data["grader_problem_id"]
+        test_cases = [(test_case["input_file"], test_case["output_file"]) for test_case in task_data["grader_test_cases"]]
+        weights = [test_case["weight"] for test_case in task_data["grader_test_cases"]]
+        options = {
+            "compute_diff": task_data["grader_compute_diffs"],
+            "treat_non_zero_as_runtime_error": task_data["treat_non_zero_as_runtime_error"],
+            "diff_max_lines": task_data["grader_diff_max_lines"],
+            "diff_context_lines": task_data["grader_diff_context_lines"],
+            "output_diff_for": [test_case["input_file"] for test_case in task_data["grader_test_cases"]
+                                if test_case["diff_shown"]]
+        }
+
+        if len(test_cases) == 0:
+            return json.dumps(
+                {"status": "error", "message": "You must provide test cases to autogenerate the grader"})
+
+        with tempfile.TemporaryDirectory() as temporary_folder_name:
+            run_file_name = 'run'
+            target_run_file = os.path.join(temporary_folder_name, run_file_name)
+
+            with open(target_run_file, "w") as f:
+                f.write(run_file_template.format(
+                    problem_id=repr(problem_id), test_cases=repr(test_cases),
+                    options=repr(options), weights=repr(weights)))
+
+            task_fs.copy_to(temporary_folder_name)
+
+    return None
+
+
+def grader_generator_tab(course, taskid, task_data, template_helper):
+    tab_id = 'tab_grader'
+    link = '<i class="fa fa-check-circle fa-fw"></i>&nbsp; Grader'
+    grader_test_cases_dump = json.dumps(task_data.get('grader_test_cases', []))
+    content = template_helper.get_custom_renderer(_BASE_RENDERER_PATH, layout=False).grader(task_data,
+                                                                                            grader_test_cases_dump)
+
+    template_helper.add_javascript('/static/grader_generator/js/grader_generator.js')
+
+    return tab_id, link, content
+
+
+def grader_footer(course, taskid, task_data, template_helper):
+    return template_helper.get_custom_renderer(_BASE_RENDERER_PATH, layout=False).grader_templates()
+
+
+def init(plugin_manager, course_factory, client, config):
+    plugin_manager.add_page(r'/static/grader_generator/(.*)', create_static_resource_page(_BASE_STATIC_FOLDER))
+
+    plugin_manager.add_hook('task_editor_tab', grader_generator_tab)
+    plugin_manager.add_hook('task_editor_footer', grader_footer)
+    plugin_manager.add_hook('task_editor_submit', on_task_editor_submit)
