+++ inginious/frontend/webapp/plugins/grader_generator/grader.html
@@ -1,4 +1,4 @@
-$def with (task_data)
+$def with (task_data, grader_test_cases_dump)
 
 $#
 $# This file is part of INGInious. See the LICENSE and the COPYRIGHTS files for
@@ -72,38 +72,56 @@ $#
     </div>
 </div>
 
-<hr>
+<hr/>
 <h4 align="center">Test cases</h4>
 <div class="jumbotron">
-<div class="row" id="grader_test_cases_header" style="display: none;">
-    <div class="form-group col-xs-12" style="text-align: center;">
-      <div class="col-xs-3"><label>Input file</label></div>
-      <div class="col-xs-3"><label>Output file</label></div>
-      <div class="col-xs-2"><label>Weight</label></div>
-      <div class="col-xs-2"><label>Differences shown</label></div>
-      <div class="col-xs-2"><label></label></div>
+    <div class="row" id="grader_test_cases_header" style="display: none;">
+        <div class="form-group col-xs-12" style="text-align: center;">
+          <div class="col-xs-3"><label>Input file</label></div>
+          <div class="col-xs-3"><label>Output file</label></div>
+          <div class="col-xs-2"><label>Weight</label></div>
+          <div class="col-xs-2"><label>Differences shown</label></div>
+          <div class="col-xs-2"><label></label></div>
+        </div>
+    </div>
+
+    <div id="grader_test_cases_container">
+    </div>
+
+    <div class="row" id="new_grader_test_case" style="display: flex;">
+        <div class="form-group col-xs-12">
+          <div class="col-xs-4">
+            <div><label for="grader_test_case_in" class="control-label">Input file</label></div>
+            <select class="form-control" id="grader_test_case_in" name="grader_test_case_in">
+            </select>
+          </div>
+          <div class="col-xs-4">
+            <div><label for="grader_test_case_out" class="control-label">Output file</label></div>
+            <select class="form-control" id="grader_test_case_out" name="grader_test_case_out">
+            </select>
+          </div>
+          <div class="col-xs-4" style="display: flex; height: 100%;">
+            <button type="button" class="btn btn-default btn-block" style="align-self: flex-end;"
+              onclick="studio_add_test_case_from_form();">Add</button>
+          </div>
+        </div>
     </div>
 </div>
 
-<div id="grader_test_cases_container">
-</div>
+<script type="text/javascript">
+    var grader_test_cases = $:{grader_test_cases_dump};
 
-<div class="row" id="new_grader_test_case" style="display: flex;">
-    <div class="form-group col-xs-12">
-      <div class="col-xs-4">
-        <div><label for="grader_test_case_in" class="control-label">Input file</label></div>
-        <select class="form-control" id="grader_test_case_in" name="grader_test_case_in">
-        </select>
-      </div>
-      <div class="col-xs-4">
-        <div><label for="grader_test_case_out" class="control-label">Output file</label></div>
-        <select class="form-control" id="grader_test_case_out" name="grader_test_case_out">
-        </select>
-      </div>
-      <div class="col-xs-4" style="display: flex; height: 100%;">
-        <button type="button" class="btn btn-default btn-block" style="align-self: flex-end;"
-          onclick="studio_add_test_case_from_form();">Add</button>
-      </div>
-    </div>
-</div>
-</div>
+    $$ (function() {
+      studio_update_grader_problems();
+
+      $$ ("#grader_problem_id").val("${task_data.get('grader_problem_id', '')}");
+      studio_load_grader_test_cases(grader_test_cases);
+
+      $$('a[data-toggle="tab"]').on('show.bs.tab', function (e) {
+        if (e.target.hash === '#tab_grader') {
+          studio_update_grader_problems();
+          studio_update_grader_files();
+        }
+      });
+    });
+</script>
