+++ inginious/frontend/webapp/plugins/grader_generator/grader_templates.html
