+++ inginious/frontend/webapp/plugins/grader_generator/run_file_template.txt
