+++ inginious/frontend/webapp/templates/course_admin/task_edit.html
@@ -1,4 +1,4 @@
-$def with (course,taskid,task_data,environments,problemdump,grader_test_cases_dump,contains_is_html,current_filetype,available_filetypes,AccessibleTime,file_list,available_languages,additional_tabs,additional_footer)
+$def with (course,taskid,task_data,environments,problemdump,contains_is_html,current_filetype,available_filetypes,AccessibleTime,file_list,available_languages,additional_tabs,additional_footer)
 
 $#
 $# This file is part of INGInious. See the LICENSE and the COPYRIGHTS files for
@@ -44,8 +44,6 @@ $var Navbar: $:NavbarF()
                 fa-cube fa-fw"></i>&nbsp; Container setup</a></li>
             <li role="presentation"><a href="#tab_subproblems" aria-controls="subproblems" role="tab" data-toggle="tab"><i class="fa
                 fa-list-ul fa-fw"></i>&nbsp; Subproblems</a></li>
-            <li role="presentation"><a href="#tab_grader" aria-controls="grader" role="tab" data-toggle="tab"><i class="fa
-                fa-check-circle fa-fw"></i>&nbsp; Grader</a></li>
             <li role="presentation"><a href="#tab_files" aria-controls="files" role="tab" data-toggle="tab"><i class="fa
                 fa-edit fa-fw"></i>&nbsp; Task files</a></li>
             $for tab_id, link, _ in additional_tabs:
@@ -62,10 +60,6 @@ $var Navbar: $:NavbarF()
 
             <div role="tabpanel" class="tab-pane" id="tab_subproblems">
                 $:include.course_admin.edit_tabs.subproblems()
-            </div>
-
-            <div role="tabpanel" class="tab-pane" id="tab_grader">
-                $:include.course_admin.edit_tabs.grader(task_data)
             </div>
 
             <div role="tabpanel" class="tab-pane" id="tab_files">
@@ -119,26 +113,9 @@ $for html in additional_footer:
 <!-- Must be out of the form -->
 $:include.course_admin.edit_tabs.subproblem_templates(contains_is_html, available_languages)
 $:include.course_admin.edit_tabs.file_modals()
-$:include.course_admin.edit_tabs.grader_templates()
 
 <!-- Init the javascript for the subproblem tab -->
 <script type="text/javascript">
     problem_data = $:{problemdump};
     $$ (function(){ studio_load(problem_data); });
-
-    grader_test_cases = $:{grader_test_cases_dump};
-
-    $$ (function() {
-      studio_update_grader_problems();
-
-      $$ ("#grader_problem_id").val("${task_data.get('grader_problem_id', '')}");
-      studio_load_grader_test_cases(grader_test_cases);
-
-      $$('a[data-toggle="tab"]').on('show.bs.tab', function (e) {
-        if (e.target.hash === '#tab_grader') {
-          studio_update_grader_problems();
-          studio_update_grader_files();
-        }
-      });
-    });
 </script>
